refactoring the task-controller

+++ src/main/java/crac/controllers/TaskController.java
@@ -9,6 +9,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -59,6 +60,7 @@ import crac.models.db.entities.Task;
 import crac.models.db.entities.Task.ArchiveTask;
 import crac.models.db.entities.Task.TaskShort;
 import crac.models.db.relation.CompetenceTaskRel;
+import crac.models.db.relation.UserCompetenceRel;
 import crac.models.db.relation.UserMaterialSubscription;
 import crac.models.db.relation.UserTaskRel;
 import crac.models.input.CompetenceTaskMapping;
@@ -209,7 +211,7 @@ public class TaskController {
 		CracUser user = userDAO.findByName(userDetails.getName());
 
 		Task task = taskDAO.findOne(id);
-		
+
 		if (task != null) {
 			try {
 				task.setTaskState(ConcreteTaskState.STARTED, taskDAO);
@@ -222,62 +224,6 @@ public class TaskController {
 		return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 
 	}
-
-	/**
-	 * Adds target task to the open-tasks of the logged-in user or changes it's
-	 * state
-	 * 
-	 * @param taskId
-	 * @return ResponseEntity
-	 */
-	/*
-	 * @RequestMapping(value = { "/{task_id}/add/{state_name}",
-	 * "/task/{task_id}/{state_name}/" }, method = RequestMethod.PUT, produces =
-	 * "application/json")
-	 * 
-	 * @ResponseBody public ResponseEntity<String> addToUser(@PathVariable(value
-	 * = "state_name") String stateName,
-	 * 
-	 * @PathVariable(value = "task_id") Long taskId) {
-	 * 
-	 * UsernamePasswordAuthenticationToken userDetails =
-	 * (UsernamePasswordAuthenticationToken) SecurityContextHolder
-	 * .getContext().getAuthentication(); CracUser user =
-	 * userDAO.findByName(userDetails.getName());
-	 * 
-	 * Task task = taskDAO.findOne(taskId);
-	 * 
-	 * if (task != null) { TaskParticipationType state =
-	 * TaskParticipationType.PARTICIPATING; if (stateName.equals("participate"))
-	 * { if (task.isJoinable()) { if (!task.isFull()) { state =
-	 * TaskParticipationType.PARTICIPATING; } else { return
-	 * JSONResponseHelper.createResponse(false, "bad_request",
-	 * ErrorCode.TASK_IS_FULL); } } else { return
-	 * JSONResponseHelper.createResponse(false, "bad_request",
-	 * ErrorCode.TASK_NOT_JOINABLE); } } else if (stateName.equals("follow")) {
-	 * if (task.isFollowable()) { state = TaskParticipationType.FOLLOWING; }
-	 * else { return JSONResponseHelper.createResponse(false, "bad_request",
-	 * ErrorCode.TASK_NOT_JOINABLE); } } else { HashMap<String, Object> meta =
-	 * new HashMap<>(); meta.put("state", stateName); return
-	 * JSONResponseHelper.createResponse(false, "bad_request",
-	 * ErrorCode.STATE_NOT_AVAILABLE, meta); }
-	 * 
-	 * UserTaskRel rel =
-	 * userTaskRelDAO.findByUserAndTaskAndParticipationTypeNot(user, task,
-	 * TaskParticipationType.LEADING);
-	 * 
-	 * if (rel == null) { rel = new UserTaskRel(); rel.setUser(user);
-	 * rel.setTask(task); rel.setParticipationType(state);
-	 * user.getTaskRelationships().add(rel); userDAO.save(user); } else {
-	 * rel.setParticipationType(state); userTaskRelDAO.save(rel); }
-	 * 
-	 * return JSONResponseHelper.successfullyAssigned(task);
-	 * 
-	 * } else { return JSONResponseHelper.createResponse(false, "bad_request",
-	 * ErrorCode.ID_NOT_FOUND); }
-	 * 
-	 * }
-	 */
 
 	@RequestMapping(value = { "/{task_id}/add/{state}",
 			"/task/{task_id}/{state}/" }, method = RequestMethod.PUT, produces = "application/json")
@@ -393,14 +339,10 @@ public class TaskController {
 		UsernamePasswordAuthenticationToken userDetails = (UsernamePasswordAuthenticationToken) SecurityContextHolder
 				.getContext().getAuthentication();
 		CracUser user = userDAO.findByName(userDetails.getName());
-		Set<UserTaskRel> trels = userTaskRelDAO.findByUserAndParticipationType(user, partType);
-		Set<ArchiveTask> tcomp = new HashSet<>();
-		for (UserTaskRel tr : trels) {
-			if (tr.getTask().getTaskState() == ConcreteTaskState.COMPLETED) {
-				tcomp.add(tr.getTask().toArchived(tr));
-			}
-		}
-		return JSONResponseHelper.createResponse(tcomp, true);
+		Set<UserTaskRel> trels = userTaskRelDAO.selectRelByUserAndParticipationTypeAndTaskState(user, partType,
+				ConcreteTaskState.COMPLETED);
+		return JSONResponseHelper.createResponse(
+				trels.stream().map(tr -> tr.getTask().toArchived(tr)).collect(Collectors.toList()), true);
 
 	}
 
@@ -423,30 +365,20 @@ public class TaskController {
 	@ResponseBody
 	public ResponseEntity<String> availableCompetences(@PathVariable(value = "task_id") Long taskId) {
 
-		Task task = taskDAO.findOne(taskId);
+		Iterable<Competence> competenceList = competenceDAO.findByDeprecatedNot(true);
 
-		Iterable<Competence> competenceList = competenceDAO.findAll();
-
-		Set<CompetenceTaskRel> competenceRels = task.getMappedCompetences();
-
-		ArrayList<Competence> found = new ArrayList<Competence>();
+		List<Competence> comps = new ArrayList<Competence>();
 
 		if (competenceList != null) {
-			for (Competence c : competenceList) {
-				boolean in = false;
-				for (CompetenceTaskRel ctr : competenceRels) {
-					if (c.getId() == ctr.getCompetence().getId()) {
-						in = true;
-					}
-				}
-				if (!in) {
-					found.add(c);
-				}
-			}
+
+			comps = StreamSupport.stream(competenceList.spliterator(), false)
+					.filter(c -> taskDAO.findOne(taskId).getMappedCompetences().stream()
+							.map(CompetenceTaskRel::getCompetence).noneMatch(uc -> uc.equals(c)))
+					.collect(Collectors.toList());
 		}
 
-		if (found.size() != 0) {
-			return JSONResponseHelper.createResponse(found, true);
+		if (comps.size() != 0) {
+			return JSONResponseHelper.createResponse(comps, true);
 		} else {
 			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.EMPTY_DATA);
 		}
@@ -721,21 +653,10 @@ public class TaskController {
 				.getContext().getAuthentication();
 		CracUser user = userDAO.findByName(userDetails.getName());
 
-		ArrayList<EvaluatedTask> tasks = new ArrayList<>();
+		Set<TaskShort> set = decider.findTasks(user, new UserFilterParameters()).stream()
+				.map(evaltask -> evaltask.getTask().toShort()).limit(numberOfTasks).collect(Collectors.toSet());
 
-		int count = 0;
-
-		for (EvaluatedTask task : decider.findTasks(user, new UserFilterParameters())) {
-
-			if (count == numberOfTasks) {
-				break;
-			}
-
-			tasks.add(task);
-			count++;
-		}
-
-		return JSONResponseHelper.createResponse(tasks, true);
+		return JSONResponseHelper.createResponse(set, true);
 
 	}
 
@@ -773,10 +694,7 @@ public class TaskController {
 				if (action.equals("overwrite")) {
 
 					Set<Material> del = new HashSet<>();
-
-					for (Material m : t.getMaterials()) {
-						del.add(m);
-					}
+					t.getMaterials().forEach(del::add);
 
 					for (Material m : del) {
 						m.getTask().getMaterials().remove(m);
@@ -939,20 +857,19 @@ public class TaskController {
 		if (st != null) {
 			if (st.isLeader(user)) {
 
-				Material delm = null;
+				Material delm = materialDAO.findOne(materialId);
 
-				for (Material m : st.getMaterials()) {
-					if (m.getId() == materialId) {
-						delm = m;
-					}
+				if (delm != null) {
+					st.getMaterials().remove(delm);
+
+					materialDAO.delete(delm);
+
+					taskDAO.save(st);
+					return JSONResponseHelper.successfullyUpdated(st);
+				} else {
+					return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+
 				}
-
-				st.getMaterials().remove(delm);
-
-				materialDAO.delete(delm);
-
-				taskDAO.save(st);
-				return JSONResponseHelper.successfullyUpdated(st);
 
 			} else {
 				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
@@ -997,7 +914,7 @@ public class TaskController {
 					String status = m.subscribable(quantity, um);
 
 					if (status.equals("OK")) {
-						
+
 						HashMap<String, Object> meta = new HashMap<>();
 						meta.put("task_id", st.getId() + "");
 						meta.put("material_id", m.getId() + "");
@@ -1052,6 +969,7 @@ public class TaskController {
 
 	/**
 	 * Set the fullfilled-variable of target subscription
+	 * 
 	 * @param subscriptionId
 	 * @param fullfilled
 	 * @return ResponseEntity
@@ -1122,69 +1040,6 @@ public class TaskController {
 
 		return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 	}
-
-	/**
-	 * Sets a single task ready to be published, only works if it's children are
-	 * ready
-	 * 
-	 * @param taskId
-	 * @return ResponseEntity
-	 */
-	/*
-	 * @RequestMapping(value = { "/{task_id}/publish/ready/single",
-	 * "/{task_id}/publish/ready/single/" }, method = RequestMethod.GET,
-	 * produces = "application/json")
-	 * 
-	 * @ResponseBody public ResponseEntity<String>
-	 * readyforPublish(@PathVariable(value = "task_id") Long taskId) { Task t =
-	 * taskDAO.findOne(taskId); UsernamePasswordAuthenticationToken userDetails
-	 * = (UsernamePasswordAuthenticationToken) SecurityContextHolder
-	 * .getContext().getAuthentication(); CracUser user =
-	 * userDAO.findByName(userDetails.getName());
-	 * 
-	 * if (t != null) { if (user.hasTaskPermissions(t)) {
-	 * 
-	 * if (t.fieldsFilled()) { if (t.childTasksReady()) {
-	 * t.setReadyToPublish(true); taskDAO.save(t); return
-	 * JSonResponseHelper.successFullyUpdated(t); } else { return
-	 * JSonResponseHelper.createResponse(false, "bad_request",
-	 * "CHILDREN_NOT_READY"); } } else { return
-	 * JSonResponseHelper.createResponse(false, "bad_request",
-	 * "TASK_NOT_READY"); }
-	 * 
-	 * } else { return JSonResponseHelper.createResponse(false, "bad_request",
-	 * "PERMISSIONS_NOT_SUFFICIENT"); } } else { return
-	 * JSonResponseHelper.idNotFound(); } }
-	 */
-
-	/**
-	 * Sets target task and all children ready to be published
-	 * 
-	 * @param taskId
-	 * @return ResponseEntity
-	 */
-	/*
-	 * @RequestMapping(value = { "/{task_id}/publish/ready/tree",
-	 * "/{task_id}/publish/ready/tree/" }, method = RequestMethod.GET, produces
-	 * = "application/json")
-	 * 
-	 * @ResponseBody public ResponseEntity<String>
-	 * forcePublish(@PathVariable(value = "task_id") Long taskId) { Task t =
-	 * taskDAO.findOne(taskId); UsernamePasswordAuthenticationToken userDetails
-	 * = (UsernamePasswordAuthenticationToken) SecurityContextHolder
-	 * .getContext().getAuthentication(); CracUser user =
-	 * userDAO.findByName(userDetails.getName());
-	 * 
-	 * if (t != null) { if (user.hasTaskPermissions(t)) { HashMap<String,
-	 * String> results = new HashMap<>(); t.readyToPublishTree(results,
-	 * taskDAO); if (results.size() == 0) { return
-	 * JSonResponseHelper.successFullyUpdated(t); } else { return
-	 * JSonResponseHelper.messageArray(results); }
-	 * 
-	 * } else { return JSonResponseHelper.createResponse(false, "bad_request",
-	 * "PERMISSIONS_NOT_SUFFICIENT"); } } else { return
-	 * JSonResponseHelper.idNotFound(); } }
-	 */
 
 	/**
 	 * Copy target task
@@ -1264,11 +1119,12 @@ public class TaskController {
 
 		Task task = taskDAO.findOne(task_id);
 		Competence competence = competenceDAO.findOne(competence_id);
+
 		if (task != null && competence != null) {
-			if (user.getCreatedTasks().contains(task) && task.getTaskState() == ConcreteTaskState.NOT_PUBLISHED
-					|| user.confirmRole("ADMIN") && task.getTaskState() == ConcreteTaskState.NOT_PUBLISHED) {
+			if (task.isLeader(user)) {
 				competenceTaskRelDAO.save(new CompetenceTaskRel(competence, task, po.getProficiencyValue(),
 						po.getImportanceValue(), po.isMandatory()));
+
 				return JSONResponseHelper.successfullyAssigned(competence);
 			} else {
 				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.RESOURCE_UNCHANGEABLE);
@@ -1508,20 +1364,24 @@ public class TaskController {
 
 		Task task = taskDAO.findOne(task_id);
 		Competence competence = competenceDAO.findOne(competence_id);
-		CompetenceTaskRel ctr = competenceTaskRelDAO.findByTaskAndCompetence(task, competence);
-		if (task != null && competence != null && ctr != null) {
-			if (user.getCreatedTasks().contains(task) && task.getTaskState() == ConcreteTaskState.NOT_PUBLISHED
-					|| user.confirmRole("ADMIN") && task.getTaskState() == ConcreteTaskState.NOT_PUBLISHED) {
-				task.getMappedCompetences().remove(ctr);
-				competence.getCompetenceTaskRels().remove(ctr);
-				competenceTaskRelDAO.delete(ctr);
-				return JSONResponseHelper.successfullyDeleted(competence);
-			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.RESOURCE_UNCHANGEABLE);
+
+		if (task != null && competence != null) {
+			CompetenceTaskRel ctr = competenceTaskRelDAO.findByTaskAndCompetence(task, competence);
+
+			if (ctr != null) {
+				if (task.isLeader(user)) {
+					task.getMappedCompetences().remove(ctr);
+					competence.getCompetenceTaskRels().remove(ctr);
+					competenceTaskRelDAO.delete(ctr);
+					return JSONResponseHelper.successfullyDeleted(competence);
+				} else {
+					return JSONResponseHelper.createResponse(false, "bad_request",
+							ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+				}
 			}
-		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
+		return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+
 	}
 
 	/**
@@ -1546,6 +1406,10 @@ public class TaskController {
 
 		PostOptions po;
 
+		UsernamePasswordAuthenticationToken userDetails = (UsernamePasswordAuthenticationToken) SecurityContextHolder
+				.getContext().getAuthentication();
+		CracUser user = userDAO.findByName(userDetails.getName());
+
 		po = mapper.readValue(json, PostOptions.class);
 
 		Task task = taskDAO.findOne(taskId);
@@ -1555,18 +1419,21 @@ public class TaskController {
 			CompetenceTaskRel ctr = competenceTaskRelDAO.findByTaskAndCompetence(task, competence);
 
 			if (ctr != null) {
-				ctr.setImportanceLevel(po.getImportanceLevel());
-				ctr.setMandatory(po.isMandatory());
-				ctr.setNeededProficiencyLevel(po.getNeededProficiencyLevel());
-				ResponseEntity<String> v = JSONResponseHelper.successfullyUpdated(ctr);
-				competenceTaskRelDAO.save(ctr);
-				return v;
-			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+				if (task.isLeader(user)) {
+					ctr.setImportanceLevel(po.getImportanceLevel());
+					ctr.setMandatory(po.isMandatory());
+					ctr.setNeededProficiencyLevel(po.getNeededProficiencyLevel());
+					ResponseEntity<String> v = JSONResponseHelper.successfullyUpdated(ctr);
+					competenceTaskRelDAO.save(ctr);
+					return v;
+				} else {
+					return JSONResponseHelper.createResponse(false, "bad_request",
+							ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+				}
 			}
-		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
+
+		return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 
 	}
 
@@ -1680,9 +1547,9 @@ public class TaskController {
 				.getContext().getAuthentication();
 		CracUser user = userDAO.findByName(userDetails.getName());
 
-		if (task.isLeader(user)) {
+		if (task != null) {
 
-			if (task != null) {
+			if (task.isLeader(user)) {
 
 				ConcreteTaskState oldState = task.getTaskState();
 
@@ -1729,10 +1596,10 @@ public class TaskController {
 				meta.put("new_state", state.toString());
 				return JSONResponseHelper.createResponse(true, meta);
 			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
 			}
 		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
 	}
 
@@ -1756,9 +1623,9 @@ public class TaskController {
 				.getContext().getAuthentication();
 		CracUser user = userDAO.findByName(userDetails.getName());
 
-		if (task.isLeader(user)) {
+		if (task != null) {
 
-			if (task != null) {
+			if (task.isLeader(user)) {
 
 				ConcreteTaskState oldState = task.getTaskState();
 
@@ -1776,10 +1643,10 @@ public class TaskController {
 
 				return JSONResponseHelper.createResponse(true, meta);
 			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
 			}
 		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
 	}
 
@@ -1837,8 +1704,8 @@ public class TaskController {
 		Task task = taskDAO.findOne(taskId);
 		List<CracUser> users = new ArrayList<>();
 
-		if (task.isLeader(user)) {
-			if (task != null) {
+		if (task != null) {
+			if (task.isLeader(user)) {
 				if (invType.equals("user")) {
 					CracUser inv = userDAO.findOne(invId);
 					if (inv != null) {
@@ -1867,10 +1734,10 @@ public class TaskController {
 				}
 				return JSONResponseHelper.successfullyCreated(users);
 			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
 			}
 		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
 
 	}
@@ -1893,8 +1760,8 @@ public class TaskController {
 		CracUser user = userDAO.findByName(userDetails.getName());
 		Task task = taskDAO.findOne(taskId);
 
-		if (task.isLeader(user)) {
-			if (task != null) {
+		if (task != null) {
+			if (task.isLeader(user)) {
 				CracGroup group = groupDAO.findOne(groupId);
 				if (group != null) {
 					if (!group.getRestrictedTasks().contains(task)) {
@@ -1907,10 +1774,10 @@ public class TaskController {
 				}
 				return JSONResponseHelper.successfullyUpdated(task);
 			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
 			}
 		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
 
 	}
@@ -1936,8 +1803,8 @@ public class TaskController {
 		CracUser user = userDAO.findByName(userDetails.getName());
 		Task task = taskDAO.findOne(taskId);
 
-		if (task.isLeader(user)) {
-			if (task != null) {
+		if (task != null) {
+			if (task.isLeader(user)) {
 
 				ObjectMapper mapper = new ObjectMapper();
 
@@ -1966,10 +1833,10 @@ public class TaskController {
 					return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 				}
 			} else {
-				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
+				return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
 			}
 		} else {
-			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.PERMISSIONS_NOT_SUFFICIENT);
+			return JSONResponseHelper.createResponse(false, "bad_request", ErrorCode.ID_NOT_FOUND);
 		}
 
 	}
